+++ src/Common/DataObjects/Pitch.ts
@@ -437,6 +437,21 @@ export class Pitch {
         ", Note: " + this.fundamentalNote + ", octave: " + this.octave.toString();
     }
 
+    /** A short representation of the note like A4 (A, octave 4), Ab5 or C#4. */
+    public ToStringShort(octaveOffset: number = 0): string {
+        let accidentalString: string = Pitch.accidentalVexflow(this.accidental);
+        if (!accidentalString) {
+            accidentalString = "";
+        }
+        const octave: number = this.octave + octaveOffset;
+        return Pitch.getNoteEnumString(this.fundamentalNote) + accidentalString + octave;
+    }
+
+    /** A shortcut getter for ToStringShort that can be useful for debugging. */
+    public get ToStringShortGet(): string {
+        return this.ToStringShort(0); // note that a getter cannot have parameters.
+    }
+
     public OperatorEquals(p2: Pitch): boolean {
         const p1: Pitch = this;
         // if (ReferenceEquals(p1, p2)) {