+++ src/VexFlowPatch/src/keysignature.js
@@ -323,9 +323,9 @@ export class KeySignature extends StaveModifier {
           glyph.setContext(this.stave.context);
           glyph.renderToStave(x);
         }
-        if (this.hidden) {
+        if (this.hidden && group) {
           // VexflowPatch: set visibility hidden, as in some systems (rendering SVG file) the key signature is rendered black even if alpha = 0
-          group?.setAttribute("visibility", "hidden"); // group is undefined for CanvasContext, e.g. in SkybottomlineCalculator
+          group.setAttribute("visibility", "hidden"); // group is undefined for CanvasContext, e.g. in SkybottomlineCalculator
         }
         this.stave.context.closeGroup();
     }